Part Three: First Submission
The dynamic activity list and search function, as well as the function to view activity details, were all unavailable. After investigation, it was found that the API server was running normally. Therefore, it was determined that the front-end JavaScript code was not executed correctly. Then, I modified three HTML files and the components.css file.