Final Revision 3
According to the assignment requirements, I changed the structure of the folder, combined three css files into one, and also deleted several js files. During this period, I found that there was a problem with my client. I added the server.js file to the front-end folder to enable the client to run normally and also fixed the CORS configuration of the api.